[UPDATE] CommentModal 구조 분리
댓글 분리
